do everything
first completed draft of the split, passes lint
